requirements.txt
# ========================================
# VTOL Drone Mission Control - Dependencies
# ========================================

# Core GUI Framework
PyQt6>=6.6.0
PyQt6-3D>=6.7.0
PyQtWebEngine>=6.6.0

# 3D Visualization & Graphics
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
open3d==0.18.0
vtk==9.3.0
matplotlib==3.8.2
plotly==5.18.0

# Numerical Computing
numpy==1.26.3
scipy==1.11.4
pandas==2.1.4

# Geospatial & Mapping
pyproj>=3.6.0
folium>=0.15.0
geopy>=2.4.0
# gdal==3.8.1  # Optional - difficult to install on Windows
opencv-python>=4.8.0
# opencv-contrib-python>=4.8.0  # Optional

# 3D Reconstruction
# pycolmap==0.5.0  # Optional - complex dependency
plyfile>=1.0.0
# pyntcloud>=0.3.0  # Optional

# ROS 2 Bridge (Install ROS 2 separately)
# Uncomment after ROS 2 installation:
# rclpy
# std_msgs
# geometry_msgs
# sensor_msgs
# nav_msgs
# mavros_msgs

# MAVLink Communication
pymavlink==2.4.41
dronekit==2.9.2

# Mission Planning
scikit-learn==1.3.2
networkx==3.2.1

# Database
sqlalchemy>=2.0.0
# sqlite3 is built into Python

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0
configparser==6.0.0

# Logging & Monitoring
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-qt==4.3.1
pytest-cov==4.1.0

# Security
cryptography==41.0.7
pyjwt==2.8.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development Tools
black==23.12.1
pylint==3.0.3
mypy==1.8.0

# Optional: Machine Learning for path optimization
# torch==2.1.2
# scikit-optimize==0.9.0
